ReactionEntityを別ファイルに分離
